+++ packages/UI/package.json
@@ -5,7 +5,7 @@
   "dependencies": {
     "@devlaunchers/models": "workspace:*",
     "@devlaunchers/utility": "workspace:*",
-    "@storybook/addon-viewport": "7.0.22",
+    "@storybook/addon-viewport": "7.0.26",
     "axios": "^0.27.2",
     "babel-loader": "8.2.5",
     "constate": "^3.3.2",
@@ -54,17 +54,17 @@
   },
   "devDependencies": {
     "@devlaunchers/eslint-config-bases": "workspace:^",
-    "@storybook/addon-a11y": "7.0.22",
-    "@storybook/addon-actions": "^7.0.22",
-    "@storybook/addon-essentials": "^7.0.22",
-    "@storybook/addon-links": "^7.0.22",
-    "@storybook/addon-mdx-gfm": "7.0.22",
+    "@storybook/addon-a11y": "7.0.26",
+    "@storybook/addon-actions": "^7.0.26",
+    "@storybook/addon-essentials": "^7.0.26",
+    "@storybook/addon-links": "^7.0.26",
+    "@storybook/addon-mdx-gfm": "7.0.26",
     "@storybook/jest": "0.1.0",
-    "@storybook/nextjs": "7.0.22",
-    "@storybook/node-logger": "^7.0.22",
-    "@storybook/react": "^7.0.22",
-    "@storybook/test-runner": "0.10.0",
-    "@storybook/testing-library": "0.1.0",
+    "@storybook/nextjs": "7.0.26",
+    "@storybook/node-logger": "^7.0.26",
+    "@storybook/react": "^7.0.26",
+    "@storybook/test-runner": "0.11.0",
+    "@storybook/testing-library": "0.2.0",
     "@testing-library/jest-dom": "5.16.4",
     "@testing-library/react": "13.3.0",
     "@testing-library/react-hooks": "8.0.1",
@@ -93,7 +93,7 @@
     "prettier": "^2.7.1",
     "rimraf": "3.0.2",
     "shell-quote": "1.7.3",
-    "storybook": "7.0.22",
+    "storybook": "7.0.26",
     "styled-components": "^5.2.1",
     "ts-jest": "28.0.7",
     "typescript": "^4.7.4"
